refactor: Ranks enum 멤버 변수 추가
이미지는 추후 변경할 예정입니다.
