--- winwin-be-domain/src/main/java/com/dpm/winwin/domain/entity/member/Ranks.java
@@ -1,19 +1,0 @@
-package com.dpm.winwin.domain.entity.member;
-
-public enum Ranks {
-    ROOKIE("루키"),
-    BEGINNER("비기너"),
-    JUNIOR("주니어"),
-    PRO("프로");
-
-
-    private final String message;
-
-    Ranks(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
