+++ winwin-be-domain/src/main/java/com/dpm/winwin/domain/entity/member/enums/TalentType.java
@@ -1,0 +1,6 @@
+package com.dpm.winwin.domain.entity.member.enums;
+
+public enum TalentType {
+    GIVE,
+    TAKE
+}
