+++ winwin-be-domain/src/main/java/com/dpm/winwin/domain/entity/talent/MemberTalent.java
@@ -2,7 +2,7 @@ package com.dpm.winwin.domain.entity.talent;
 
 import com.dpm.winwin.domain.entity.BaseEntity;
 import com.dpm.winwin.domain.entity.member.Member;
-import com.dpm.winwin.domain.entity.talent.enums.TalentType;
+import com.dpm.winwin.domain.entity.member.enums.TalentType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
