--- winwin-be-domain/src/main/java/com/dpm/winwin/domain/entity/talent/enums/TalentType.java
@@ -1,6 +1,0 @@
-package com.dpm.winwin.domain.entity.talent.enums;
-
-public enum TalentType {
-    GIVE,
-    TAKE
-}
